requirements.txt
python-dotenv
langchain-google-generativeai
langchain_community
langchain
faiss-cpu
streamlit